requirements.txt
dm_control
Pillow
pygame
PyQt6
pyrealsense2
pure-python-adb
tyro
zmq
termcolor
pyquaternion
pyyaml
rospkg
pexpect
mujoco==2.3.7
dm_control==1.0.14
opencv-python
matplotlib
einops
packaging
h5py
ipython
tyro==0.8.8
huggingface_hub==0.25.2
click
timm
open_clip_torch>=2.20.0
torch
torchvision
numpy
tqdm
